+++ CHANGELOG.md
@@ -1,23 +1,33 @@
+# CHANGELOG
+
+## 1.1.0 (December 13th, 2018)
+
+Features:
+
+- Added "dryrun" mode, to allow scheduler to output API call details out to the log without actually firing the API calls
+- Removed need to specify instance zone in configuration
+- General tidy up of the code
+
 ## 1.0.3 (July 4th, 2017)
 
 Update:
 
-  - Updated third-party cron library to fix duplicate API call issue.
+- Updated third-party cron library to fix duplicate API call issue.
 
 ## 1.0.2 (February 7th, 2017)
 
 Defect fix:
 
-  - Unable to use non-default configuration file name
+- Unable to use non-default configuration file name
 
 ## 1.0.1 (November 14th, 2016)
 
 Features:
 
-  - Added ability to generate dynamic date strings for insertion in to API parameters
+- Added ability to generate dynamic date strings for insertion in to API parameters
 
 ## 1.0.0 (November 12th, 2016)
 
 Features:
 
-  - Initial Release
+- Initial Release